third_party/blink/web_tests/platform/generic/http/tests/security/mixedContent/redirect-https-to-http-script-in-iframe-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-redirect-https-to-http-script.html' was loaded over HTTPS, but requested an insecure script 'http://example.test:8080/security/mixedContent/resources/script.js'. This content should also be served over HTTPS.
This test loads a secure iframe that loads an insecure script (but with a tricky redirect). We should trigger a mixed content callback because an active network attacker can end up controling the script.
